Moved parameter constraints: kafe2/fit/_base -> kafe2/core
