+++ kafe2/core/constraint.py
