+++ kafe2/fit/representation/constraint/yaml_drepr.py
@@ -1,8 +1,8 @@
 from .._base import DReprError
-from .._yaml_base import YamlWriterMixin, YamlReaderMixin, YamlReaderException
+from .._yaml_base import YamlWriterMixin, YamlReaderMixin
 from ._base import ConstraintDReprBase
 from .. import _AVAILABLE_REPRESENTATIONS
-from ....fit._base.constraint import GaussianSimpleParameterConstraint, GaussianMatrixParameterConstraint
+from kafe2.core.constraint import GaussianSimpleParameterConstraint, GaussianMatrixParameterConstraint
 
 __all__ = ['ConstraintYamlWriter', 'ConstraintYamlReader']
